ref(replay): Add jsdoc to all replay modules
This is one more step of https://github.com/getsentry/sentry-javascript/issues/6323.

Note that this also renames/moves a few things to make more sense.
