+++ packages/replay/src/session/saveSession.ts
@@ -1,6 +1,9 @@
 import { REPLAY_SESSION_KEY, WINDOW } from '../constants';
 import type { Session } from '../types';
 
+/**
+ * Save a session to session storage.
+ */
 export function saveSession(session: Session): void {
   const hasSessionStorage = 'sessionStorage' in WINDOW;
   if (!hasSessionStorage) {
