+++ packages/replay/src/util/createBreadcrumb.ts
@@ -2,6 +2,9 @@ import type { Breadcrumb } from '@sentry/types';
 
 type RequiredProperties = 'category' | 'message';
 
+/**
+ * Create a breadcrumb for a replay.
+ */
 export function createBreadcrumb(
   breadcrumb: Pick<Breadcrumb, RequiredProperties> & Partial<Omit<Breadcrumb, RequiredProperties>>,
 ): Breadcrumb {
