+++ packages/replay/src/util/createReplayEnvelope.ts
@@ -1,6 +1,10 @@
 import { DsnComponents, ReplayEnvelope, ReplayEvent, ReplayRecordingData } from '@sentry/types';
 import { createEnvelope, createEventEnvelopeHeaders, getSdkMetadataForEnvelopeHeader } from '@sentry/utils';
 
+/**
+ * Create a replay envelope ready to be sent.
+ * This includes both the replay event, as well as the recording data.
+ */
 export function createReplayEnvelope(
   replayEvent: ReplayEvent,
   recordingData: ReplayRecordingData,
